Stop re-loading values inside KCoreConfigSkeleton::save().
The extra load of values in KCoreConfigSkeleton is not documented anywhere
that it happens, and in normal circumstances re-loading new values isn't
expecterd during a save operation.

Update various mentions of readConfig/writeConfig to read/save.

Update documentation to match the new names of functions.

Rename writeConfig() to save() and usrWriteConfig() to usrSave()

Rename functions to match the new naming conventions.  Also create old deprecated
instances of both functions to ensure that old code continues to operate
correctly.  Also make save() non-virtual, for the same reasons read() is
now non-virtual.

Rename usrReadConfig to usrRead, to be consistent with the new name for readConfig.

Rename usrReadConfig to usrRead, and mark the former as deprecated.  To maintain
compatibility, usrRead still calls usrReadConfig in its default implementation.
usrReadConfig remains empty.

REVIEW: 117010