Fix reading of entries for language/country combinations
This fixes a regression introduced in
988f09bb051dca0437ecec431ee44ed5b4a560d8.

The mentioned commit ensures that if the locale is e.g. "de_DE" the
entry "de" will be used. But this breaks if there is a translation
for another country. E.g. for "de_CH" it would also pick the "de"
entry.

This change now operates on both just the language code and the locale.
If an entry with the language code is present it will be picked. If
another entry with the exact locale is found it will be overwritten.
Modifiers are not supported as this is currently missing in QLocale.

REVIEW: 118692