Fix build on Visual C++ 2010.
Visual C++ 2010 throws an internal compiler error trying to compile
kconfigtest, since 5f4dc2973f. I couldn't figure out a non-intrusive
code tweak that would work around the ICE, so I had to skip compiling
this particular initialization code on this particular compiler version,
and skip running the test that relies on it (testEnums).

The dummy=42 entry is so that the config group isn't empty,
which makes testGroupCopyTo and testReparent still run and pass,
so we don't need to skip them.

REVIEW:118852