Don't upgrate config file with upd file from kde4.
This patch is necessary because:
When we use kf5 + kde4 application, kconf_update which launchs at the start when we launch kde.
But it will migrate some config file, for example it will show that we need to migrate konversation
so it will create a konversationrc in .config/

But when we launch konversation there is a kdelibs4migrator which wants to migrate settings and config in .config
but it shows a konversationrc in .config so it will never migrate and we will lose all settings.

So we can force to remove all .upd in kf5 but it will not fix problem during migration or when we have kde4 application
install in same directory as kf5.

So now I force for each upd file to have a "Version=5" so kconf_update (kf5) will migrate just kf5 upd file and it will fix my bugs.

REVIEW: 121797
CHANGELOG: Now kconf_update doesn't process upd file from kde4. We need to add "Version=5" in top of the upd file otherwise it will be skipped.