Revert the SIC in commit 915976c1238be811f169eab1b02f7e8dad6410e0.
On further reflection, introducing a SIC to fix the bug is not ideal.  I'm
working on something that will address the underlying bug, but I didn't have
it in time for 5.7.

The versioning behaviour is left, as I do want it to exist for the future.
This only reverts the mandatory versioning of files to avoid the compatibility
break.

CHANGELOG: Undo 915976c123, which never saw a release, enough to avoid the SIC.