Ensure group is unescaped properly in kconf_update.
During a kconf_update run, an invalid group name may be treated as correct
even though the name failed to unescape.  This leads the group name to be
a null character, which will fail.

If the unescape failed, return that failure status instead.  This should have
no impact, as the previous result would have been wrong anyways.  Now a
more useful diagnostic will be returned instead.

Update the unit tests to ensure this issue is tested in the future.

Found in Coverity issue 258087.

REVIEW: 126556