Don't be fatal on File field not being properly parsed
Summary:
Fallback on trying the same filename as the actual file.
There were some regressions due to the former matching expression:
string(REGEX REPLACE "^(.*\n)?File=([^\n]+kcfg).*\n.*$" "\\2"  _kcfg_FILENAME "${_contents}")

This one would allow for the File field to be the kcfgc file (which doesn't
make sense).
With this patch we're not reproducing this behaviour, but we're allowing for the
things that aren't building to work. After all, most kcfg and kcfgc files share
the same name.

Reviewers: #frameworks, davidedmundson

Reviewed By: davidedmundson

Subscribers: aacid, kfunk

Differential Revision: https://phabricator.kde.org/D3226