Generate an instance with KSharedConfig::Ptr for singleton and arg
Summary:
In case a kcfg with arg="true" was used and singleton the static
instance method only accepted a QString config name. This made it
impossible to combine a singleton config with an already existing and
open KSharedConfig::Ptr.

With this change an overloaded instance method is added which takes a
KSharedConfig::Ptr as argument. The private ctor, though, only takes a
KSharedConfig::Ptr and the instance method taking a QString argument
uses KSharedConfig::openConfig on the config file name.

This provides full API compatibility and at the same time allows to use
KSharedConfig in addition to the arg name based variant.

Test Plan:
kconfigcompiler tests still pass and a config with singleton
and arg="true" generates the code as I need it

Reviewers: #frameworks

Differential Revision: https://phabricator.kde.org/D3386