KConfig: stop exporting and installing KConfigBackend.
Summary:
It can't possibly have been used anywhere, because it's not in the KConfig
API anywhere. The intended way to use this API was by providing a plugin
that would derive from KConfigBackend but the plugin loading code
in KConfigBackend::create has been disabled since before KF 5.0.

The reason I want to stop exporting this class is to be able to
optimize it (e.g. the QDateTime in it is completely unused but leads
to data races due to tzset)

Test Plan: Note, I'll rename it to _p.h if we agree, it would just have made this diff too big.

Reviewers: mdawson

Reviewed By: mdawson

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D4604