Fix the result of KDesktopFile::sortOrder()
Summary:
KDesktopFile::sortOrder() returns the value of SortOrder key
as a string whithout parsing it as a list.
But according to Desktop Entry Specification
https://standards.freedesktop.org/desktop-entry-spec/desktop-entry-spec-latest.html
it's type is "string(s)", i.e. the same type as the type of
Actions and MimeType keys, and thus it should be read the same way.

Reviewers: #frameworks, dfaure

Reviewed By: dfaure

Subscribers: dfaure, #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D8689