Don't look for /etc/kderc every single time
Summary:
Every time we open a configuration file, we are checking if this file is there.
Since it's a sysadmin setting I'd say it's fine to assume that it's not appearing
and disappearing.
Also we are not supporting the case of the file changing during runtime or so.

Reviewers: #frameworks, mpyne, dfaure

Reviewed By: mpyne, dfaure

Subscribers: dfaure, mpyne

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D8871