kcfg_compiler now documents valid inputs for its 'Color' type
Summary:
it wasn't exactly clear what the inputs for a Color type may be. this is
now explained along with special type descriptions. it isn't perfectly
ideal here since it is a bit hard to find, but better be hard to find
than not documented at all.

from a quick glance at the code indeed anything goes that QColor can
construct from a QString. there is also custom regex code in the compiler
that allows construction from r,g,b,a via the appropriate ctor of qcolor.

Reviewers: broulik

Reviewed By: broulik

Subscribers: broulik, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D15576