Add mechanism to notify other clients of config changes over DBus
Summary:
Intention is not to create a registry like system, but to replace
KDElibs4Support::KGlobalSettings and to replace other system settingss
-> some app communication in a more generic way.

writeEntry gains an additional flag Notify which if set, will notify
clients of what has actually changed when we sync.

Rationale to put this into KConfig was so that we could have everything
batched and sychronised to the file sync and to get the fine detailed
exposure of what has actually changed which we don't get with a file
watcher.

Default behaviour remains identical without any broadcast messages.

As it is a new dependency it is purely optional and anything referencing
DBus is not in the public API. Our deployment on platforms without DBus
tend to be standalone applications anyway.

Test Plan: Attached unit test

Reviewers: broulik, dfaure

Reviewed By: broulik, dfaure

Subscribers: dfaure, broulik, zzag, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D13034