kconfig_compiler: new kcfgc args HeaderExtension & SourceExtension
Summary:
When using kconfig_compiler generated sources in projects where the used
file extensions for C++ files are not ".h" and/or ".cpp", the include for
the generated header looks a bit alien to the project, as well as the
generated source file if one looks closer at it.

This code adds new optional flags HeaderExtension & SourceExtension which
can be used to control the file extensions used for the generated files.

Test Plan:
All unit tests and the new test_fileextensions pass, existing projects using
kcfg without & with the new flags build fine.

Reviewers: #frameworks, apol

Reviewed By: apol

Subscribers: apol, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D19565