KCONFIG_ADD_KCFG_FILES macro: ensure a change of File= in kcfg is picked up
Summary:
The build configuration depends on the value of the "File=" entry in the
kcfg file, as this file name is used in the build instructions.
So if the name is changed, cmake would need to be rerun.
Adding the kcfgc file to CMAKE_CONFIGURE_DEPENDS makes cmake know about that
dependency.
While this will also result in a reconfiguarion if non-File entries are
edited, this should not happen too often, so the extra costs outweighs the
unexpected and confusing behaviour due to outdated build instructions
in case the File= entry is changed.

Test Plan:
Before this change renaming a kcfg file before did not trigger a rerun of cmake,
resulting in outdated builds instructions and unexpected behaviour.
With this change cmake is rerun once the kcgc file is edited, so build
instructions are always up-to-date.

Reviewers: #frameworks, apol

Reviewed By: apol

Subscribers: apol, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D19567