[KConfig] port away from deprecated methods in Qt 5.14
Summary:
In kconf_update, the ctime usage used to be about metadata change time
(buff.st_ctime, before it got ported to the misnamed created()).
I ported it to birthTime, because I think
date of birth is a more useful way to identify a file than
date of permission change which doesn't really matter to us.
But in practice, I can't help but wonder if mtime alone wouldn't be
enough.

For the QStringLiteral("%%1").arg(i) bit, I tested it in tst_qstring,
the first % is left untouched.

Test Plan: make && ctest

Reviewers: mdawson, arichardson, vkrause

Reviewed By: vkrause

Subscribers: pino, arojas, mlaurent, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D23815