Fix writeFlags with KConfigCompilerSignallingItem
Summary:
KConfigCompilerSignallingItem both inherits KConfigSkeletonItem and
internally is powered by a separate KConfigSkeletonItem

The generated code calls setWriteFlags on the outer
KConfigCompilerSignallingItem instance, but the real writing is
performed by the internal version. We need to set the flags in the right
place.

Ideally we would do this in an overload of KConfigSkeletonItem, but
given we can't, I've shadowed the method. This isn't pretty, but given
the docs say it should generally only be used from auto generated code,
should be fine.

Test Plan: Used in workspace KCM

Reviewers: ervin

Reviewed By: ervin

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D25791