Fix windows being inappropriately maximized on launch
When a window is closed while maximized, we write a special string to
the config file so that it gets restored in its maximized state. But we
don't ever delete that thing when the window is un-maximized and closed
again, causing the window to always be maximized when launched.

BUG: 426813
FIXED-IN: 5.78