Fix restoring window size when closed while maximized
The maximization string was being saved to the config file including
the connected screens, but the code to read it was not, and was also
looking for the old incorrect format. This could have resulted in old
config values being used forever, and the window always being opened
in a maximized state.

BUG: 430521
FIXED-IN: 5.79