Fix window positioning regression due to incorrect fallback in fallback
We have a fallback string that gets searched for in the KCM, but that
fallback itself should not have a fallback, or else we always use that
second fallback and the window gets positioned incorrectly the first
time it's opened. So let's replace the second fallback with -1.

BUG: 432661
FIXED-IN: 5.79
(cherry picked from commit 97accd14d732f3eda1c25bb58290cb9311040798)