Allow kreadconfig/kwriteconfig to access the root group
The default group if "--group" is not specified is "KDE".
Trying to specify a null group name causes an assert in KConfigGroup::groupImpl().
Explicitly detect and give an error message for a null group name.
Mention in the help text that "<default>" can be used to specify the root group -
this string comes from KConfigGroupPrivate::name().

Add the missing parser.addHelpOption() for kreadconfig.
Update the help strings.
Apply clang-format changes.
    
I18N: