Make singleton teardown work with Qt6 as well
In Qt6 the Q_GLOBAL_STATIC will already report to be null while it is in
the process of being deleted, we therefore cannot access it anymore from
destruction code path as we did before.

This problem is hit for example by the Breeze style, making all 6 based
widgets applications crash on exit without this.