Add KConfigSkeletonChangeNotifyingItem for KConfigCompilerSignallingItem
The prefix of the name KConfigCompilerSignallingItem is not matching
the other class names and possibly was missed to be adapted, as
the other name KConfigSkeletonChangeNotifyingItem with a prefix
matching the other classes can be found in the API dox instead.

While so far the class was not used in the public API of generated code
fixing the generation of code for ItemAccessors=true with signalling items
will leak this class onto the consumer side, so better give it a nice name
and also move member data to a pimpl class.

# Conflicts:
#	src/core/kcoreconfigskeleton.h