Change "Actual Size" shortcut's text to "Zoom to Actual Size"
In general actions need to start with verbs so that the user knows what
will happen. Changing "Actual Size" to "Zoom to Actual Size" makes the
action compliant with this rule, and also makes it clear that this is a
zoom action, so anywhere the user searches for the word "zoom", they
will find it.