kconfig_compiler: fix generation of bit masks enum items for signals
In case we need to generate signals for the notification of keys, an
anonymous enum is created with bit mask items for each key; since there
is a limit of bits available, the check used to be:
  unsigned val = 1 << parseResult.signalList.size();
Since "1" is int by default, and parseResult.signalList is a QList (and
thus size() is int), the whole operation is on int, and for a shift of
more than 31 (i.e. more than 31 keys to generate signals for) the result
is undefined behaviour. This UB manifested itself in two ways:
(a) kconfig_compiler works, however all the enum items after the 31st
    have a value of 0x0
(b) kconfig_compiler exits with the message:
    "Too many signals to create unique bit masks"

Hence, change the way the enum values are generated, also according to
the fact that KConfigCompilerSignallingItem uses quint64 for the user
data (which represents the signals triggered):
- use quint64 for the value of each enum (left-shifted each iteration)
  instead of unsigned int
- tweak the "too many signals" condition by checking the bits that fit
  in quint64, removing the UB
Note that, due to the limitation of the implementation in
KConfigCompilerSignallingItem, it will not be possible to generate more
than 64 notification signals.

Add a unit test for this with a single kcfg file with 64 entries,
requesting to generate signals; in addition to generating it (checking
that (b) is fixed), also add a very small unit test checking the enum
value for the 64th key is not 0x0 but the value expected (checking that
(a) is fixed).