kconfig_compiler: switch away from bit mask for signal enum values
kconfig_compiler used to have a bitmask enum for the values identifying
notification signals of properties. This approach was implemented with
commit 6d3a4405fc2723f5796e845247b6b0eb0448216e possibly as efficient
way to collect which signals were emitted, so it is easier to emit them
again later on. The big drawback is that it is possible to store max 64
keys, as that is the bit count of quint64.

To solve this properly, switch away from the bit mask way: the enum
value of each signal is a simple incrementing integer. This is possible
only one value for a key is signalled by KConfigCompilerSignallingItem
at a time, so the slot that reacts to all the instances of
KConfigCompilerSignallingItem does not need to account for more than one
at the same time. The other change needed is the collection of emitted
signals for the usrSave() implementation: switch from uint to QSet, so
it can store all the needed values (making sure to #include <QSet> only
where needed).

The only potential incompatibility is due to the changed values of the
"signal*" enum items, as that anynonymous enum is public; hopefully
there are no users of those in application code...

Update the tests accordingly:
- drop test_64_entries, no more needed now, as the limitation is gone
- update the reference baseline for the changed output

Updates/fixes commit 6c0acf3be392414bbdfbb7473aa842ee59d65f3d.