Apply existing  QScreen::name() workaround for Windows to new code
e5fb03b6e915b176ae8804ae7317ff1977848487 added a workaround for a Qt
bug on Windows that makes QScreen::name() return garbage.

21e02655a78b5d241c35fd934f7665564420327c added new un-worked-around
usages of QScreen::name() which effectively re-introduces the bug that
the old code was fixing on Windows.

This commit re-applies the workaround to the new code using a compiler
macro.

BUG: 429943
FIXED-IN: 5.101