Sort connector names for multi-screen size/position keys
A string including the connector names is used in the config file key for
storing per-screen-arrangement size and position data, which means we
need this string to be consistent for the same screen arrangement. But
connector order is non-deterministic. We need to sort the list to keep a
consistent order and avoid losing multi-screen size and position data.

CCBUG: 460260
(cherry picked from commit 3f84ef821d76cd7c646d4328b4627eb70dae4ab3)